--- a/tracer/build/_build/Build.Steps.cs
+++ b/tracer/build/_build/Build.Steps.cs
@@ -63,7 +63,7 @@ partial class Build
 
     const string LibDdwafVersion = "1.15.1";
 
-    string[] OlderLibDdwafVersions = new[] { "1.3.0", "1.10.0" };
+    string[] OlderLibDdwafVersions = new[] { "1.3.0", "1.10.0", "1.14.0" };
 
     AbsolutePath LibDdwafDirectory(string libDdwafVersion = null) => (NugetPackageDirectory ?? RootDirectory / "packages") / $"libddwaf.{libDdwafVersion ?? LibDdwafVersion}";