--- a/tracer/src/Datadog.Trace/AppSec/Waf/NativeBindings/NativeLibrary.cs
+++ b/tracer/src/Datadog.Trace/AppSec/Waf/NativeBindings/NativeLibrary.cs
@@ -68,6 +68,23 @@ namespace Datadog.Trace.AppSec.Waf.NativeBindings
             return handle != IntPtr.Zero;
         }
 
+        internal static bool CloseLibrary(IntPtr library)
+        {
+            if (library != IntPtr.Zero)
+            {
+                Log.Error("Trying to close a library with a null pointer");
+                return false;
+            }
+
+            if (isPosixLike)
+            {
+                var result = NonWindows.dddlclose(library);
+                return result == 0;
+            }
+
+            return FreeLibrary(library);
+        }
+
         private static IntPtr LoadWindowsLibrary(string libraryPath)
         {
             var handle = LoadLibrary(libraryPath);
@@ -138,6 +155,9 @@ namespace Datadog.Trace.AppSec.Waf.NativeBindings
         [DllImport("kernel32.dll", SetLastError = true)]
         private static extern IntPtr LoadLibrary(string dllToLoad);
 
+        [DllImport("kernel32.dll", SetLastError = true)]
+        private static extern bool FreeLibrary(IntPtr libraryToFree);
+
         [DllImport("Kernel32.dll", SetLastError = true)]
         private static extern IntPtr GetProcAddress(IntPtr hModule, string lpProcName);
 
@@ -161,6 +181,9 @@ namespace Datadog.Trace.AppSec.Waf.NativeBindings
 #pragma warning disable SA1300 // Element should begin with upper-case letter
             [DllImport("Datadog.Tracer.Native")]
             internal static extern IntPtr dddlopen(string fileName, int flags);
+
+            [DllImport("Datadog.Tracer.Native")]
+            internal static extern int dddlclose(IntPtr library);
 
             [DllImport("Datadog.Tracer.Native")]
             internal static extern IntPtr dddlerror();