--- a/tracer/src/Datadog.Trace/Datadog.Trace.csproj
+++ b/tracer/src/Datadog.Trace/Datadog.Trace.csproj
@@ -62,7 +62,6 @@
     <None Remove="AppSec\Waf\ConfigFiles\rule-set.json" />
     <EmbeddedResource Include="AppSec\Waf\ConfigFiles\rule-set.json" />
     <None Remove="AppSec\Waf\ConfigFiles\apisecurity-config.json" />
-    <EmbeddedResource Include="AppSec\Waf\ConfigFiles\apisecurity-config.json" />
   </ItemGroup>
 
   <!-- Below this point is reserved for vendor items -->