--- a/tracer/test/Datadog.Trace.Security.Unit.Tests/WafErrorsTests.cs
+++ b/tracer/test/Datadog.Trace.Security.Unit.Tests/WafErrorsTests.cs
@@ -40,7 +40,7 @@ namespace Datadog.Trace.Security.Unit.Tests
             waf.Should().NotBeNull();
             initResult.Success.Should().BeTrue();
             initResult.FailedToLoadRules.Should().Be(0);
-            initResult.LoadedRules.Should().Be(152);
+            initResult.LoadedRules.Should().Be(153);
             initResult.Errors.Should().BeEmpty();
             initResult.HasErrors.Should().BeFalse();
             initResult.ErrorMessage.Should().BeNullOrEmpty();