--- a/tracer/test/snapshots/Security.ApiSecurity.AspNetCore5.ApiSecOff.__url=_dataapi_empty-model_body={-property_expectedStatusCode=NoContent_containsAttack=False.verified.txt
+++ b/tracer/test/snapshots/Security.ApiSecurity.AspNetCore5.ApiSecOff.__url=_dataapi_empty-model_body={-property_expectedStatusCode=NoContent_containsAttack=False.verified.txt
@@ -1,4 +1,4 @@
-[
+﻿[
   {
     TraceId: Id_1,
     SpanId: Id_2,