--- a/tracer/tools/Samples.Transport.UnixDomainSocket/Samples.Transport.UnixDomainSocket.csproj
+++ b/tracer/tools/Samples.Transport.UnixDomainSocket/Samples.Transport.UnixDomainSocket.csproj
@@ -29,7 +29,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.14.2" />
+    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.14.0" />
   </ItemGroup>
 
   <ItemGroup>