## Summary of changes

Update waf 1.5.1 and rules 1.10.0

The upgrade to WAF 1.5.1 introduces the concept of "ephemeral addresses". The are designed for streaming scenarios where most of the WAF context will be unchanged but the streamed here will need to be repeatedly updated.

This feature is described in more detail here:
https://github.com/DataDog/libddwaf/blob/1.15.0/UPGRADING.md#ephemeral-addresses