@robertpi 2023-12-15
> we might want to update the function `CheckVersionCompatibility` in `WafConfigurator` as well as there's a breaking change for previous tracer versions, there's also `WafCompatibilityTests` but might not be necessary to add an additional test for 2.43

fixed in 0b9ad25e7b9c01930daf80f0c6c3ced029d69c5c, though it took a bit of refactoring, as only missing exports where tested