@robertpi 2023-12-20
```
```mermaid
gantt
    title Throughput Linux x64 (ASM) (Total requests) 
    dateFormat  X
    axisFormat %s
    section Baseline
    This commit (74bedb) (7.430M)   : 0, 7429704
    master (7.351M)   : 0, 7351334
    benchmarks/2.9.0 (7.675M)   : 0, 7675468

    section No attack
    This commit (74bedb) (1.757M)   : crit ,0, 1757315
    master (1.994M)   : 0, 1994109
    benchmarks/2.9.0 (3.152M)   : 0, 3151744

    section Attack
    This commit (74bedb) (1.428M)   : crit ,0, 1427512
    master (1.571M)   : 0, 1571157
    benchmarks/2.9.0 (2.514M)   : 0, 2513990

    section Blocking
    This commit (74bedb) (3.154M)   : 0, 3153988
    master (3.289M)   : 0, 3289021

    section IAST default
    This commit (74bedb) (6.649M)   : 0, 6649435
    master (6.435M)   : 0, 6434844

    section IAST full
    This commit (74bedb) (5.759M)   : 0, 5759211
    master (5.774M)   : 0, 5773825

    section Base vuln
    This commit (74bedb) (0.928M)   : 0, 928031
    master (0.941M)   : 0, 941308

    section IAST vuln
    This commit (74bedb) (0.872M)   : 0, 871514
    master (0.873M)   : 0, 873056

```